+++ src/wayland/seat/mod.rs
@@ -49,7 +49,7 @@
 //! This module further defines the `"cursor_image"` role, that is assigned to surfaces used by clients
 //! to change the cursor icon.
 
-mod keyboard;
+pub(crate) mod keyboard;
 mod pointer;
 mod touch;