+++ wlcs_anvil/src/main_loop.rs
@@ -18,6 +18,7 @@ use smithay::{
     },
     utils::IsAlive,
     wayland::{
+        input_method::InputMethodSeat,
         output::{Mode, Output, PhysicalProperties},
         seat::{ButtonEvent, CursorImageStatus, MotionEvent},
         SERIAL_COUNTER as SCOUNTER,
@@ -25,7 +26,7 @@ use smithay::{
 };
 
 use anvil::{
-    drawing::{draw_cursor, draw_dnd_icon},
+    drawing::{draw_cursor, draw_dnd_icon, draw_input_popup_surface},
     render::render_output,
     state::Backend,
     AnvilState, CalloopData, ClientState,
@@ -115,6 +116,22 @@ pub fn run(channel: Channel<WlcsEvent>) {
                     ));
                 }
             }
+
+            // draw input method square if any
+            let input_method = state.seat.input_method().unwrap();
+            let rectangle = input_method.coordinates();
+            input_method.with_surface(|surface| {
+                elements.push(
+                    draw_input_popup_surface(
+                        surface.clone(),
+                        (
+                            rectangle.loc.x + rectangle.size.w,
+                            (rectangle.loc.y + rectangle.size.h),
+                        ),
+                    )
+                    .into(),
+                );
+            });
 
             // draw the cursor as relevant
             // reset the cursor if the surface is no longer alive